requirements.txt
python-binance==1.0.34
python-dotenv==1.2.1
requests==2.32.5
